+++ NEWS.md
@@ -4,7 +4,8 @@
   * `bucket` now depends on `slate` function
 * Modified `prop2prob` function
   * Return p-values as a sorted `data.table`
-  * Let user select `method` for p-value adjustment
+  * Now lets user select `method` for p-value adjustment
+  * New `prompt` argument turns off big data prompt
   * Fix pass by reference bug in `linRcpp`
 * New `abstract` function
   * Combines two `propr` objects into one